+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 * Moved Sanitizer Provider interfaces to lucid-support
+* Removed force* methods
+* Added Processor level validation hook
+* Added Email Processor
 
 ## v0.1.0 (2022-09-07)
 * Built initial Sanitizer structure